+++ package.json
@@ -12,9 +12,11 @@
     "i18next": "^21.6.11",
     "i18next-browser-languagedetector": "^6.1.3",
     "i18next-http-backend": "^1.3.2",
+    "js-cookie": "^3.0.1",
     "react": "^17.0.2",
     "react-bootstrap": "^2.1.2",
     "react-dom": "^17.0.2",
+    "react-flag-icon-css": "^1.0.25",
     "react-i18next": "^11.15.3",
     "react-icons": "^4.3.1",
     "react-redux": "^7.2.6",
