+++ public/assets/locales/en/translation.json
@@ -2,5 +2,5 @@
     "app_title": "Title",
     "language": "Language",
     "welcome_message": "Welcome",
-    "days_since_release": "it's been {{number_of_days}} days since this video was released"
+    "days_since_release": "it's been {{number_of_days}} days since this i18n was started"
   }
