+++ src/Pages/UserPage.scss
@@ -1,4 +1,7 @@
-.userPage{
-    display: flex;
-    justify-content: space-evenly;
+.userPage {
+  display: flex;
+  justify-content: space-evenly;
+  .flag-icon:before {
+    padding-right: 25px;
+  }
 }
