+++ src/i18n.js
@@ -17,7 +17,7 @@ i18n
     backend:{
       loadPath:'/assets/locales/{{lng}}/translation.json',
     },
-    react:{useSuspense:false },
+    
   });
 
 export default i18n;
