+++ src/index.js
@@ -1,19 +1,24 @@
-import React from 'react';
+import React, {Suspense} from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from "react-redux";
-import 'bootstrap/dist/js/bootstrap.min.js';
-import 'bootstrap/dist/css/bootstrap.min.css';
 import App from './App';
 import configureStore from "./redux/store/configureStore";
 import "./i18n";
-const store = configureStore();
 
+const store = configureStore();
+const loadingMarkup = (
+  <div className='py-4 text-center'>
+    <h2>Loading ...</h2>
+  </div>
+)
 ReactDOM.render(
+  <Suspense fallback={loadingMarkup}>
   <React.StrictMode>
     <Provider store={store} >
       <App />
       </Provider>
-  </React.StrictMode>,
+  </React.StrictMode>
+  </Suspense>,
   document.getElementById('root')
 );
